work in progress, largely untested
